+++ plugins/backend/backstage-plugin-argo-cd-backend/src/service/router.ts
@@ -4,6 +4,7 @@ import express from 'express';
 import Router from 'express-promise-router';
 import { Logger } from 'winston';
 import { ArgoService } from './argocd.service';
+import { InstanceConfig } from './types';
 
 export interface RouterOptions {
   logger: Logger;
@@ -26,10 +27,58 @@ export function createRouter({
     config.getOptionalString('argocd.password') ?? 'argocdPassword';
   const argoSvc = new ArgoService(argoUserName, argoPassword, config, logger);
 
+  const findArgoInstance = (argoInstanceName: string) => {
+    const argoInstanceArray = argoSvc.getArgoInstanceArray();
+    const foundArgoInstance = argoInstanceArray.find(
+      argoInstance => argoInstance.name === argoInstanceName,
+    );
+    return foundArgoInstance;
+  };
+
+  async function findMatchedArgoInstanceToken(
+    matchedArgoInstance: InstanceConfig,
+  ) {
+    let token: string;
+    if (!matchedArgoInstance.token) {
+      token = await argoSvc.getArgoToken(matchedArgoInstance);
+    } else {
+      token = matchedArgoInstance.token;
+    }
+    return token;
+  }
+
   router.get('/find/name/:argoAppName', async (request, response) => {
     const argoAppName = request.params.argoAppName;
     response.send(await argoSvc.findArgoApp({ name: argoAppName }));
   });
+
+  router.get(
+    '/argoInstance/:argoInstanceName/applications/name/:argoAppName/revisions/:revisionID/metadata',
+    async (request, response) => {
+      const revisionID: string = request.params.revisionID;
+      const argoInstanceName: string = request.params.argoInstanceName;
+      const argoAppName = request.params.argoAppName;
+      logger.info(`Getting info on ${argoAppName}`);
+      logger.info(`Getting app ${argoAppName} on ${argoInstanceName}`);
+      const matchedArgoInstance = findArgoInstance(argoInstanceName);
+      if (matchedArgoInstance === undefined) {
+        return response.status(500).send({
+          status: 'failed',
+          message: 'cannot find an argo instance to match this cluster',
+        });
+      }
+      const token: string = await findMatchedArgoInstanceToken(
+        matchedArgoInstance,
+      );
+      const resp = await argoSvc.getRevisionData(
+        matchedArgoInstance.url,
+        { name: argoAppName },
+        token,
+        revisionID,
+      );
+      return response.send(resp);
+    },
+  );
 
   router.get(
     '/argoInstance/:argoInstanceName/applications/name/:argoAppName',
@@ -38,22 +87,16 @@ export function createRouter({
       const argoAppName = request.params.argoAppName;
       logger.info(`Getting info on ${argoAppName}`);
       logger.info(`Getting app ${argoAppName} on ${argoInstanceName}`);
-      const argoInstanceArray = argoSvc.getArgoInstanceArray();
-      const matchedArgoInstance = argoInstanceArray.find(
-        argoInstance => argoInstance.name === argoInstanceName,
-      );
+      const matchedArgoInstance = findArgoInstance(argoInstanceName);
       if (matchedArgoInstance === undefined) {
         return response.status(500).send({
           status: 'failed',
           message: 'cannot find an argo instance to match this cluster',
         });
       }
-      let token: string;
-      if (!matchedArgoInstance.token) {
-        token = await argoSvc.getArgoToken(matchedArgoInstance);
-      } else {
-        token = matchedArgoInstance.token;
-      }
+      const token: string = await findMatchedArgoInstanceToken(
+        matchedArgoInstance,
+      );
       const resp = await argoSvc.getArgoAppData(
         matchedArgoInstance.url,
         matchedArgoInstance.name,
@@ -75,22 +118,16 @@ export function createRouter({
       logger.info(
         `Getting apps for selector ${argoAppSelector} on ${argoInstanceName}`,
       );
-      const argoInstanceArray = argoSvc.getArgoInstanceArray();
-      const matchedArgoInstance = argoInstanceArray.find(
-        argoInstance => argoInstance.name === argoInstanceName,
-      );
+      const matchedArgoInstance = findArgoInstance(argoInstanceName);
       if (matchedArgoInstance === undefined) {
         return response.status(500).send({
           status: 'failed',
           message: 'cannot find an argo instance to match this cluster',
         });
       }
-      let token: string;
-      if (!matchedArgoInstance.token) {
-        token = await argoSvc.getArgoToken(matchedArgoInstance);
-      } else {
-        token = matchedArgoInstance.token;
-      }
+      const token: string = await findMatchedArgoInstanceToken(
+        matchedArgoInstance,
+      );
       const resp = await argoSvc.getArgoAppData(
         matchedArgoInstance.url,
         matchedArgoInstance.name,
@@ -108,10 +145,7 @@ export function createRouter({
     const labelValue = request.body.labelValue as string;
     const sourceRepo = request.body.sourceRepo;
     const sourcePath = request.body.sourcePath;
-    const argoInstanceArray = argoSvc.getArgoInstanceArray();
-    const matchedArgoInstance = argoInstanceArray.find(
-      argoInstance => argoInstance.name === argoInstanceName,
-    );
+    const matchedArgoInstance = findArgoInstance(argoInstanceName);
     if (matchedArgoInstance === undefined) {
       return response.status(500).send({
         status: 'failed',