+++ plugins/backend/catalog-backend-module-okta/src/providers/OktaGroupEntityProvider.ts
@@ -31,7 +31,9 @@ import {
 import { AccountConfig } from '../types';
 import { groupEntityFromOktaGroup } from './groupEntityFromOktaGroup';
 import { getAccountConfig } from './accountConfig';
-import { assertError } from '@backstage/errors';
+import { isError } from '@backstage/errors';
+import { getOktaGroups } from './getOktaGroups';
+import { getParentGroup } from './getParentGroup';
 
 /**
  * Provides entities from Okta Group service.
@@ -41,18 +43,31 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
   private readonly userNamingStrategy: UserNamingStrategy;
   private readonly groupFilter: string | undefined;
   private readonly orgUrl: string;
-  private readonly parentGroupField: string | undefined;
+  private hierarchyConfig: { parentKey: string; key?: string } | undefined;
 
   static fromConfig(
     config: Config,
     options: {
       logger: winston.Logger;
-      parentGroupField?: string;
       namingStrategy?: GroupNamingStrategies | GroupNamingStrategy;
       userNamingStrategy?: UserNamingStrategies | UserNamingStrategy;
+      /*
+       * @deprecated, please use hierarchyConfig.parentKey
+       */
+      parentGroupField?: string;
+      hierarchyConfig?: {
+        parentKey: string;
+        key?: string;
+      };
     },
   ) {
     const accountConfig = getAccountConfig(config);
+
+    if (options.parentGroupField && !options.hierarchyConfig?.parentKey) {
+      options.hierarchyConfig = {
+        parentKey: options.parentGroupField,
+      };
+    }
 
     return new OktaGroupEntityProvider(accountConfig, options);
   }
@@ -61,19 +76,22 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
     accountConfig: AccountConfig,
     options: {
       logger: winston.Logger;
-      parentGroupField?: string;
       namingStrategy?: GroupNamingStrategies | GroupNamingStrategy;
       userNamingStrategy?: UserNamingStrategies | UserNamingStrategy;
+      hierarchyConfig?: {
+        parentKey: string;
+        key?: string;
+      };
     },
   ) {
     super([accountConfig], options);
-    this.parentGroupField = options.parentGroupField;
     this.namingStrategy = groupNamingStrategyFactory(options.namingStrategy);
     this.userNamingStrategy = userNamingStrategyFactory(
       options.userNamingStrategy,
     );
     this.orgUrl = accountConfig.orgUrl;
     this.groupFilter = accountConfig.groupFilter;
+    this.hierarchyConfig = options.hierarchyConfig;
   }
 
   getProviderName(): string {
@@ -92,34 +110,54 @@ export class OktaGroupEntityProvider extends OktaEntityProvider {
 
     const defaultAnnotations = await this.buildDefaultAnnotations();
 
-    await client.listGroups({ search: this.groupFilter }).each(async group => {
-      const members: string[] = [];
-      await group.listUsers().each(user => {
-        try {
-          const userName = this.userNamingStrategy(user);
-          members.push(userName);
-        } catch (e) {
-          assertError(e);
-          this.logger.warn(`failed to add user to group: ${e.message}`);
-        }
-      });
-
-      try {
-        const groupEntity = groupEntityFromOktaGroup(
-          group,
-          this.namingStrategy,
-          {
-            annotations: defaultAnnotations,
-            members,
-            parentGroupField: this.parentGroupField,
-          },
-        );
-        groupResources.push(groupEntity);
-      } catch (e) {
-        assertError(e);
-        this.logger.warn(`failed to add group: ${e.message}`);
-      }
+    const oktaGroups = await getOktaGroups({
+      client,
+      groupFilter: this.groupFilter,
+      groupNamingStrategy: this.namingStrategy,
+      key: this.hierarchyConfig?.key,
+      logger: this.logger,
     });
+
+    await Promise.allSettled(
+      Object.entries(oktaGroups).map(async ([_, group]) => {
+        const members: string[] = [];
+        await group.listUsers().each(user => {
+          try {
+            const userName = this.userNamingStrategy(user);
+            members.push(userName);
+          } catch (e: unknown) {
+            this.logger.warn(
+              `failed to add user to group: ${
+                isError(e) ? e.message : 'unknown error'
+              }`,
+            );
+          }
+        });
+
+        const parentGroup = getParentGroup({
+          parentKey: this.hierarchyConfig?.parentKey,
+          group,
+          oktaGroups,
+        });
+
+        try {
+          const groupEntity = groupEntityFromOktaGroup(
+            group,
+            this.namingStrategy,
+            parentGroup,
+            {
+              annotations: defaultAnnotations,
+              members,
+            },
+          );
+          groupResources.push(groupEntity);
+        } catch (e) {
+          this.logger.warn(
+            `failed to add group: ${isError(e) ? e.message : 'unknown error'}`,
+          );
+        }
+      }),
+    );
 
     await this.connection.applyMutation({
       type: 'full',